获取用户firmid

+++ ipo-service-api/src/main/java/com/yrdce/ipo/modules/sys/service/DisplayService.java
@@ -4,6 +4,6 @@ public interface DisplayService {
 
 	public String userInfo(String userid);
 
-	public Integer display(String sid, String money1);
+	public Long display(String sid, String money1);
 
 }
